🔥 Disable 'prefer-destructuring' rule by default
This rule can create some nasty-looking code, and does not always
work as expected. For example, there is no obvious way for the
following snippet to be rewritten such that it would satisfy this
rule without the use of an intermediate variable:

```
alice.lastName = bob.lastName
```

Because this line of code is, by all measure, harmless, electing
to disable the rule in which it is in violation.